Flytt ut hentAvslagsperioderMedBegrunnelser i utilfunksjon
